#286: Moved the character-related models to a new sub-folder "character"
Gathering stuff a bit.
