+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.spec.ts
@@ -14,8 +14,8 @@ import {HTML_LINE_RETURN} from './skill-effects.mapper';
 import {Equipment} from '../../model/equipment/equipment.model';
 import {EquipmentsService} from '../../services/equipments.service';
 import {EQUIPMENTS_TEST_DATA} from '../../model/equipment/equipment.model.spec';
-import {CHARACTER_TEST_DATA} from '../../model/character.model.spec';
-import {Character} from '../../model/character.model';
+import {CHARACTER_TEST_DATA} from '../../model/character/character.model.spec';
+import {Character} from '../../model/character/character.model';
 import {CharactersService} from '../../services/characters.service';
 import {SkillsServiceMock} from '../../services/skills.service.spec';
 import {EquipmentsServiceMock} from '../../services/equipments.service.spec';
