+++ src/app/ffbe/mappers/equipment-mapper.spec.ts
@@ -5,8 +5,8 @@ import {EquipmentMapper} from './equipment-mapper';
 import {Equipment} from '../model/equipment/equipment.model';
 import {CharactersServiceMock} from '../services/characters.service.spec';
 import {CharactersService} from '../services/characters.service';
-import {CHARACTER_TEST_DATA} from '../model/character.model.spec';
-import {Character} from '../model/character.model';
+import {CHARACTER_TEST_DATA} from '../model/character/character.model.spec';
+import {Character} from '../model/character/character.model';
 import {Caracteristiques} from '../model/caracteristiques.model';
 import {PASSIVE_SKILLS_TEST_DATA} from '../model/skill.model.spec';
 import {Skill} from '../model/skill.model';
