+++ src/app/ffbe/model/character/character-entry-stats.model.ts
