+++ src/app/ffbe/model/character/character.model.ts
