+++ src/app/ffbe/services/characters.service.spec.ts
@@ -2,9 +2,9 @@ import {inject, TestBed} from '@angular/core/testing';
 import {Observable, of} from 'rxjs';
 import {CharactersService} from './characters.service';
 import {DataMiningClientService} from './data-mining-client.service';
-import {CHARACTER_TEST_DATA} from '../model/character.model.spec';
+import {CHARACTER_TEST_DATA} from '../model/character/character.model.spec';
 import {SkillsService} from './skills.service';
-import {Character} from '../model/character.model';
+import {Character} from '../model/character/character.model';
 import {LimitBurstsService} from './limit-bursts.service';
 import {SkillsServiceMock} from './skills.service.spec';
 import {ABILITY_SKILLS_TEST_DATA, MAGIC_SKILLS_TEST_DATA, PASSIVE_SKILLS_TEST_DATA} from '../model/skill.model.spec';
